+++ smartbreak/src/screens/goals/dicas.js
@@ -2,7 +2,7 @@ import { StatusBar } from 'expo-status-bar';
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
 
-export default function TestGoal() {
+export default function Dicas({ navigation }) {
 
   
   return (
